#2514 Refactor and add automated tests
#2514 Support importing multiline fields
