+++ de.metas.adempiere.adempiere/client/src/test/resources/multiplelines.csv
@@ -1,0 +1,10 @@
+G0006	eMail	offen	2008-02-02	2009-04-02	"29.4.2009 - Bla bal bla test bla test bla
+
+15.04.2009 - test, dfsd, test bla bla
+
+31.03.2009
+- Bla bal bla test bla test bla 
+- Bla bal bla test bla test bla
+- Bla bal bla test bla test bla
+- Infos test bla test bla"	70
+G0007	eMail	offen	2008-03-18	2013-01-16	Präsentation vor Ort	80
