+++ de.metas.adempiere.adempiere/client/src/test/resources/regularlines.csv
@@ -1,0 +1,3 @@
+G0007	eMail	offen	2008-03-18	2013-01-16	Präsentation vor Ort	80
+G0007	eMail	offen	2008-03-18	2013-01-16	Präsentation vor Ort	90
+G0007	eMail	offen	2008-03-18	2013-01-16	Präsentation vor Ort	100
